+++ src/numerics/APRReconstruction.hpp
@@ -29,11 +29,11 @@ public:
 
         for (uint64_t level = apr_iterator.level_min(); level <= apr_iterator.level_max(); ++level) {
             //loop over the resolutions of the structure
-            const unsigned int x_num_ =  apr.spatial_index_x_max(level);
-            const unsigned int z_num_ = apr.spatial_index_z_max(level);
+            const int64_t x_num_ =  apr.spatial_index_x_max(level);
+            const int64_t z_num_ = apr.spatial_index_z_max(level);
 
-            const unsigned int x_num_min_ = 0;
-            const unsigned int z_num_min_ = 0;
+            const int64_t x_num_min_ = 0;
+            const int64_t z_num_min_ = 0;
 
             const float step_size = pow(2,apr_iterator.level_max() - level);
 
@@ -59,14 +59,14 @@ public:
                 const int offset_max_dim2 = std::min((int) img.x_num, (int) (dim2 + step_size));
                 const int offset_max_dim3 = std::min((int) img.z_num, (int) (dim3 + step_size));
 
-                for (int q = dim3; q < offset_max_dim3; ++q) {
+                for (int64_t q = dim3; q < offset_max_dim3; ++q) {
 
-                    for (int k = dim2; k < offset_max_dim2; ++k) {
+                    for (int64_t k = dim2; k < offset_max_dim2; ++k) {
     #ifdef HAVE_OPENMP
 	#pragma omp simd
 #endif
-                        for (int i = dim1; i < offset_max_dim1; ++i) {
-                            img.mesh[(size_t)i + (k) * img.y_num + q * img.y_num * img.x_num] = temp_int;
+                        for (int64_t i = dim1; i < offset_max_dim1; ++i) {
+                            img.mesh[i + (k) * img.y_num + q * img.y_num * img.x_num] = temp_int;
                         }
                     }
                 }
@@ -178,9 +178,9 @@ public:
         //
 
 
-        const int z_num = input.z_num;
-        const int x_num = input.x_num;
-        const int y_num = input.y_num;
+        const int64_t z_num = input.z_num;
+        const int64_t x_num = input.x_num;
+        const int64_t y_num = input.y_num;
 
         std::vector<T> temp_vec;
         temp_vec.resize(y_num,0);
@@ -189,7 +189,7 @@ public:
         offset_vec.resize(y_num,0);
 
 
-        int i, k, index;
+        int64_t i, k, index;
         float counter, temp, divisor,offset;
 
         //need to introduce an offset max to make the algorithm still work, and it also makes sense.
@@ -202,7 +202,7 @@ public:
 #ifdef HAVE_OPENMP
 	#pragma omp parallel for default(shared) private(i,k,counter,temp,index,divisor,offset) firstprivate(temp_vec,offset_vec)
 #endif
-        for(int j = 0;j < z_num;j++){
+        for(int64_t j = 0;j < z_num;j++){
             for(i = 0;i < x_num;i++){
 
                 index = j*x_num*y_num + i*y_num;
@@ -210,7 +210,7 @@ public:
                 //first update the fixed length scale
                 for (k = 0; k < y_num;k++){
 
-                    offset_vec[k] = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)index + k])/scale),(T)offset_max);
+                    offset_vec[k] = std::min((T)floor(pow(2,d_max- offset_img.mesh[index + k])/scale),(T)offset_max);
 
                 }
 
@@ -218,11 +218,11 @@ public:
                 //first pass over and calculate cumsum
                 temp = 0;
                 for (k = 0; k < y_num;k++){
-                    temp += input.mesh[(size_t)index + k];
+                    temp += input.mesh[index + k];
                     temp_vec[k] = temp;
                 }
 
-                input.mesh[(size_t)index] = 0;
+                input.mesh[index] = 0;
                 //handling boundary conditions (LHS)
                 for (k = 1; k <= (offset_max+1);k++){
                     divisor = 2*offset_vec[k] + 1;
@@ -230,7 +230,7 @@ public:
 
                     if(k <= (offset+1)){
                         //emulate the bc
-                        input.mesh[(size_t)index + k] = -temp_vec[0]/divisor;
+                        input.mesh[index + k] = -temp_vec[0]/divisor;
                     }
 
                 }
@@ -240,7 +240,7 @@ public:
                     divisor = 2*offset_vec[k] + 1;
                     offset = offset_vec[k];
                     if(k >= (offset+1)){
-                        input.mesh[(size_t)index + k] = -temp_vec[k - offset - 1]/divisor;
+                        input.mesh[index + k] = -temp_vec[k - offset - 1]/divisor;
                     }
 
                 }
@@ -251,7 +251,7 @@ public:
                     divisor = 2*offset_vec[k] + 1;
                     offset = offset_vec[k];
                     if(k < (y_num - offset)) {
-                        input.mesh[(size_t)index + k] += temp_vec[k + offset] / divisor;
+                        input.mesh[index + k] += temp_vec[k + offset] / divisor;
                     }
                 }
 
@@ -266,9 +266,9 @@ public:
                     if(k >= (y_num - offset)){
                         counter = k - (y_num-offset)+1;
 
-                        input.mesh[(size_t)index + k]*= divisor;
-                        input.mesh[(size_t)index + k]+= temp_vec[y_num-1];
-                        input.mesh[(size_t)index + k]*= 1.0/(divisor - counter);
+                        input.mesh[index + k]*= divisor;
+                        input.mesh[index + k]+= temp_vec[y_num-1];
+                        input.mesh[index + k]*= 1.0/(divisor - counter);
 
                     }
 
@@ -281,7 +281,7 @@ public:
                     offset = offset_vec[k];
 
                     if(k < (offset + 1)){
-                        input.mesh[(size_t)index + k] *= divisor/(1.0*k + offset);
+                        input.mesh[index + k] *= divisor/(1.0*k + offset);
                     }
 
                 }
@@ -289,13 +289,12 @@ public:
                 //end point boundary condition
                 divisor = 2*offset_vec[0] + 1;
                 offset = offset_vec[0];
-                input.mesh[(size_t)index] *= divisor/(offset+1);
+                input.mesh[index] *= divisor/(offset+1);
             }
         }
-
-
-
     }
+
+
     template<typename T>
     void calc_sat_adaptive_x(MeshData<T>& input,MeshData<uint8_t>& offset_img,float scale_in,unsigned int offset_max_in,const unsigned int d_max){
         //
@@ -303,18 +302,18 @@ public:
         //
         //
 
-        const int z_num = input.z_num;
-        const int x_num = input.x_num;
-        const int y_num = input.y_num;
+        const int64_t z_num = input.z_num;
+        const int64_t x_num = input.x_num;
+        const int64_t y_num = input.y_num;
 
         unsigned int offset_max = offset_max_in;
 
         std::vector<T> temp_vec;
         temp_vec.resize(y_num*(2*offset_max + 2),0);
 
-        int i,k;
+        int64_t i,k;
         float temp;
-        int index_modulo, previous_modulo, current_index, jxnumynum, offset,forward_modulo,backward_modulo;
+        int64_t index_modulo, previous_modulo, current_index, jxnumynum, offset,forward_modulo,backward_modulo;
 
         const float scale = scale_in;
         //const unsigned int d_max = this->depth_max();
@@ -332,13 +331,13 @@ public:
 
             for(k = 0; k < y_num ; k++){
                 // std::copy ?
-                temp_vec[k] = input.mesh[(size_t)jxnumynum + k];
+                temp_vec[k] = input.mesh[jxnumynum + k];
             }
 
 
             for(i = 1; i < 2 * offset_max + 1; i++) {
                 for(k = 0; k < y_num; k++) {
-                    temp_vec[i*y_num + k] = input.mesh[(size_t)jxnumynum + i*y_num + k] + temp_vec[(i-1)*y_num + k];
+                    temp_vec[i*y_num + k] = input.mesh[jxnumynum + i*y_num + k] + temp_vec[(i-1)*y_num + k];
                 }
             }
 
@@ -347,9 +346,9 @@ public:
 
             for(i = 0; i < offset_max + 1; i++){
                 for(k = 0; k < y_num; k++) {
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)jxnumynum + i * y_num + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[jxnumynum + i * y_num + k])/scale),(T)offset_max);
                     if(i < (offset + 1)) {
-                        input.mesh[(size_t)jxnumynum + i * y_num + k] = (temp_vec[(i + offset) * y_num + k]) / (i + offset + 1);
+                        input.mesh[jxnumynum + i * y_num + k] = (temp_vec[(i + offset) * y_num + k]) / (i + offset + 1);
                     }
                 }
             }
@@ -364,14 +363,14 @@ public:
                 for(k = 0; k < y_num; k++) {
 
 
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)jxnumynum + i * y_num + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[jxnumynum + i * y_num + k])/scale),(T)offset_max);
 
                     if((i >= offset_max + 1) & (i < (x_num - offset_max))) {
                         // update the buffers
 
                         index_modulo = (i + offset_max) % (2 * offset_max + 2);
                         previous_modulo = (i + offset_max - 1) % (2 * offset_max + 2);
-                        temp = input.mesh[(size_t)jxnumynum + (i + offset_max) * y_num + k] + temp_vec[previous_modulo * y_num + k];
+                        temp = input.mesh[jxnumynum + (i + offset_max) * y_num + k] + temp_vec[previous_modulo * y_num + k];
                         temp_vec[index_modulo * y_num + k] = temp;
 
                     }
@@ -381,7 +380,7 @@ public:
                         // calculate the positions in the buffers
                         forward_modulo = (i + offset) % (2 * offset_max + 2);
                         backward_modulo = (i - offset - 1) % (2 * offset_max + 2);
-                        input.mesh[(size_t)jxnumynum + i * y_num + k] = (temp_vec[forward_modulo * y_num + k] - temp_vec[backward_modulo * y_num + k]) /
+                        input.mesh[jxnumynum + i * y_num + k] = (temp_vec[forward_modulo * y_num + k] - temp_vec[backward_modulo * y_num + k]) /
                                                                 (2 * offset + 1);
 
                     }
@@ -395,14 +394,14 @@ public:
 
                 for(k = 0; k < y_num; k++){
 
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)jxnumynum + i * y_num + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[jxnumynum + i * y_num + k])/scale),(T)offset_max);
 
                     if(i >= (x_num - offset)){
                         // calculate the positions in the buffers
                         backward_modulo  = (i - offset - 1) % (2 * offset_max + 2); //maybe the top and the bottom different
                         forward_modulo = (x_num - 1) % (2 * offset_max + 2); //reached the end so need to use that
 
-                        input.mesh[(size_t)jxnumynum + i * y_num + k] = (temp_vec[forward_modulo * y_num + k] -
+                        input.mesh[jxnumynum + i * y_num + k] = (temp_vec[forward_modulo * y_num + k] -
                                                                  temp_vec[backward_modulo * y_num + k]) /
                                                                 (x_num - i + offset);
                     }
@@ -419,14 +418,14 @@ public:
 
         // The same, but in place
 
-        const int z_num = input.z_num;
-        const int x_num = input.x_num;
-        const int y_num = input.y_num;
+        const int64_t z_num = input.z_num;
+        const int64_t x_num = input.x_num;
+        const int64_t y_num = input.y_num;
 
-        int j,k;
+        int64_t j,k;
         float temp;
-        int index_modulo, previous_modulo, current_index, iynum,forward_modulo,backward_modulo,offset;
-        int xnumynum = x_num * y_num;
+        int64_t index_modulo, previous_modulo, current_index, iynum,forward_modulo,backward_modulo,offset;
+        int64_t xnumynum = x_num * y_num;
 
         const int offset_max = offset_max_in;
         const float scale = scale_in;
@@ -447,22 +446,22 @@ public:
 
             for(k = 0; k < y_num ; k++){
                 // std::copy ?
-                temp_vec[k] = input.mesh[(size_t)iynum + k];
+                temp_vec[k] = input.mesh[iynum + k];
             }
 
             //(updated z)
             for(j = 1; j < 2 * offset_max+ 1; j++) {
                 for(k = 0; k < y_num; k++) {
-                    temp_vec[j*y_num + k] = input.mesh[(size_t)j * xnumynum + iynum + k] + temp_vec[(j-1)*y_num + k];
+                    temp_vec[j*y_num + k] = input.mesh[j * xnumynum + iynum + k] + temp_vec[(j-1)*y_num + k];
                 }
             }
 
             // LHS boundary (updated)
             for(j = 0; j < offset_max + 1; j++){
                 for(k = 0; k < y_num; k++) {
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)j * xnumynum + iynum + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[j * xnumynum + iynum + k])/scale),(T)offset_max);
                     if(i < (offset + 1)) {
-                        input.mesh[(size_t)j * xnumynum + iynum + k] = (temp_vec[(j + offset) * y_num + k]) / (j + offset + 1);
+                        input.mesh[j * xnumynum + iynum + k] = (temp_vec[(j + offset) * y_num + k]) / (j + offset + 1);
                     }
                 }
             }
@@ -472,7 +471,7 @@ public:
 
                 for(k = 0; k < y_num; k++) {
 
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)j * xnumynum + iynum + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[j * xnumynum + iynum + k])/scale),(T)offset_max);
 
                     //update the buffer
                     if((j >= offset_max + 1) & (j < (z_num - offset_max))) {
@@ -481,7 +480,7 @@ public:
                         previous_modulo = (j + offset_max - 1) % (2 * offset_max + 2);
 
                         // the current cumsum
-                        temp = input.mesh[(size_t)(j + offset_max) * xnumynum + iynum + k] + temp_vec[previous_modulo*y_num + k];
+                        temp = input.mesh[(j + offset_max) * xnumynum + iynum + k] + temp_vec[previous_modulo*y_num + k];
                         temp_vec[index_modulo*y_num + k] = temp;
                     }
 
@@ -490,7 +489,7 @@ public:
                         forward_modulo = (j + offset) % (2 * offset_max + 2);
                         backward_modulo = (j - offset - 1) % (2 * offset_max + 2);
 
-                        input.mesh[(size_t)j * xnumynum + iynum + k] =
+                        input.mesh[j * xnumynum + iynum + k] =
                                 (temp_vec[forward_modulo * y_num + k] - temp_vec[backward_modulo * y_num + k]) /
                                 (2 * offset + 1);
 
@@ -503,14 +502,14 @@ public:
             for(j = z_num - offset_max; j < z_num; j++){
                 for(k = 0; k < y_num; k++){
 
-                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[(size_t)j * xnumynum + iynum + k])/scale),(T)offset_max);
+                    offset = std::min((T)floor(pow(2,d_max- offset_img.mesh[j * xnumynum + iynum + k])/scale),(T)offset_max);
 
                     if(j >= (z_num - offset)){
                         //calculate the buffer offsets
                         backward_modulo  = (j - offset - 1) % (2 * offset_max + 2); //maybe the top and the bottom different
                         forward_modulo = (z_num - 1) % (2 * offset_max + 2); //reached the end so need to use that
 
-                        input.mesh[(size_t)j * xnumynum + iynum + k] = (temp_vec[forward_modulo*y_num + k] -
+                        input.mesh[j * xnumynum + iynum + k] = (temp_vec[forward_modulo*y_num + k] -
                                                                 temp_vec[backward_modulo*y_num + k]) / (z_num - j + offset);
 
                     }
